Added benchmark function. Changed eigenToMatlab so that it doesn't zero-initialize arrays before setting them. Moved eigenToMatlab, matlabToEigen, etc. to drakeUtil so that they're not copied all over the place.
