+++ systems/controllers/QPControllermex.cpp
@@ -13,6 +13,7 @@
 #include <Eigen/StdVector>
 #include <limits>
 #include <cmath>
+#include "drakeUtil.h"
 
 //#define TEST_FAST_QP
 //#define USE_MATRIX_INVERSION_LEMMA
