+++ systems/controllers/bodyMotionControlmex.cpp
@@ -1,7 +1,8 @@
 /* 
  * A simple PD control block for regulating a body pose given a desired position, velocity, and acceleration.   
  */
- #include "controlUtil.h"
+#include "controlUtil.h"
+#include "drakeUtil.h"
 
 struct BodyMotionControlData {
   RigidBodyManipulator* r;
